ObjectMappingクラスを外に出す
