+++ Engine/Tool/BoolToolElement.cs
@@ -44,7 +44,7 @@ namespace Altseed2
         /// <param name="source">バインディング対象</param>
         /// <param name="objectMapping"></param>
         /// <returns></returns>
-        public static BoolToolElement Create(object source, ToolElementManager.ObjectMapping objectMapping)
+        public static BoolToolElement Create(object source, ObjectMapping objectMapping)
         {
             return new BoolToolElement(objectMapping.Name, source, objectMapping.PropertyName);
         }
