+++ Engine/Tool/EnumToolElement.cs
@@ -46,7 +46,7 @@ namespace Altseed2
         /// <param name="source"></param>
         /// <param name="objectMapping"></param>
         /// <returns></returns>
-        public static EnumToolElement Create(object source, ToolElementManager.ObjectMapping objectMapping)
+        public static EnumToolElement Create(object source, ObjectMapping objectMapping)
         {
             return new EnumToolElement(objectMapping.Name, source, objectMapping.PropertyName);
         }
