+++ Engine/Tool/GroupToolElement.cs
@@ -66,7 +66,7 @@ namespace Altseed2
         /// <param name="source"></param>
         /// <param name="objectMapping"></param>
         /// <returns></returns>
-        public static GroupToolElement Create(object source, ToolElementManager.ObjectMapping objectMapping)
+        public static GroupToolElement Create(object source, ObjectMapping objectMapping)
         {
             return new GroupToolElement(objectMapping.Name, source, objectMapping.PropertyName);
         }
