+++ Engine/Tool/IntToolElement.cs
@@ -68,7 +68,7 @@ namespace Altseed2
         /// <param name="source"></param>
         /// <param name="objectMapping"></param>
         /// <returns></returns>
-        public static IntToolElement Create(object source, ToolElementManager.ObjectMapping objectMapping)
+        public static IntToolElement Create(object source, ObjectMapping objectMapping)
         {
             var speed = objectMapping.Options.ContainsKey("speed") ? (float)objectMapping.Options["speed"] : 1;
             var min = objectMapping.Options.ContainsKey("min") ? (int)objectMapping.Options["min"] : -100;
