+++ Engine/Tool/UserToolElement.cs
@@ -66,7 +66,7 @@ namespace Altseed2
         /// <param name="source"></param>
         /// <param name="objectMapping"></param>
         /// <returns></returns>
-        public static UserToolElement Create(object source, ToolElementManager.ObjectMapping objectMapping)
+        public static UserToolElement Create(object source, ObjectMapping objectMapping)
         {
             return new UserToolElement(objectMapping.Name, source, objectMapping.PropertyName);
         }
